requirements.txt
Flask==2.0.1
google-api-python-client==2.27.0
google-auth==1.35.0
google-auth-oauthlib==0.4.6
Werkzeug==2.0.1
gunicorn==20.1.0